requirements.txt
charset-normalizer==2.1.1
docker==6.0.1
idna==3.4
numpy==1.21.6
packaging==21.3
pandas==1.3.5
psycopg2==2.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
websocket-client==1.4.2
wincertstore==0.2
